+++ ricaun.Nuke/Components/IRelease.cs
@@ -44,7 +44,7 @@ namespace ricaun.Nuke.Components
             var fileNameVersion = GetReleaseFileNameVersion(project.Name, version);
             var ProjectDirectory = ReleaseDirectory / fileNameVersion;
 
-            var nupkgs = Globbing.GlobFiles(ContentDirectory, "**/*.nupkg");
+            var nupkgs = Globbing.GlobFiles(ContentDirectory, "**/*.*nupkg");
             if (nupkgs.Count > 0)
             {
                 nupkgs.ForEach(file => AbsolutePathExtensions.CopyToDirectory(file, ProjectDirectory));